// lib_ui/src/app/services/models/book-detail-response.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface BookDetailResponse {
  authors?: string;
  avgRating?: number;
  bookStatus?: 'AVAILABLE' | 'BORROWED' | 'LOST';
  borrowedCount?: number;
  categoryName?: string;
  detail?: string;
  id?: number;
  name?: string;
  ratingCount?: number;
}